implemented the item menu
KBInventory uses now the NSNotificationCenter to inform other objects about the change of the selected Item
KBInventory now has two properties for the selected tool and the selected item. It also has methods for an easy access to all tools an items that currently are in the bag of the player. IMPORTANT: items that the player holds, aren't in the bag!
KBItem has now a Property ItemType, to distinguish between Items and Tools
enum ItemType added
many bugfixes for the KBGameMenuLayer
the user can now select items with the ingmae menu